pyproject.toml
[project]
name = "fastapi-freecodeacademy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.1",
    "fastapi[standard]>=0.115.12",
    "passlib[argon2,bcrypt]>=1.7.4",
    "psycopg2>=2.9.10",
    "pydantic-settings>=2.9.1",
    "pyjwt[crypto]>=2.10.1",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.2",
]
